ENH: Revert changes for pipelines to save/load files
The definition of a pipeline is now to be MRML nodes only for input and
output. This is to keep the pipelines simple. Saving and loading from
files is going to be manual now, but will be handled in the future by
the (currently non-existent) case iterator module and a (also currently
non-existent) script that will be generated by the pipeline creator for
command line usage.